added cut functionality
